Create consensus subroutines and use in distributed opt methods
