+++ libensemble/gen_funcs/persistent_n_agent.py
@@ -3,14 +3,13 @@ import numpy.linalg as la
 import scipy.sparse as spp
 
 from libensemble.message_numbers import STOP_TAG, PERSIS_STOP, FINISHED_PERSISTENT_GEN_TAG
-from libensemble.tools.gen_support import sendrecv_mgr_worker_msg
+from libensemble.tools.consensus_subroutines import print_final_score, get_grad, get_neighbor_vals
 
 def n_agent(H, persis_info, gen_specs, libE_info):
     """ Gradient sliding. Coordinates with alloc to do local and distributed 
         (i.e., gradient of consensus step) calculations.
     """
-    # TODO: Allow early termination by checking tago
-    # WHAT DO I MEAN WITH ABOVE?
+    # TODO: Allow early termination by checking tag (what do I mean here?)
     tag = None
     ub = gen_specs['user']['ub']
     lb = gen_specs['user']['lb']
@@ -78,64 +77,3 @@ def n_agent(H, persis_info, gen_specs, libE_info):
             print('[{}]: {}'.format(local_gen_id, x_k), flush=True)
 
     return None, persis_info, FINISHED_PERSISTENT_GEN_TAG
-
-def get_grad(x, f_i_idxs, gen_specs, libE_info):
-    l = len(f_i_idxs)
-    H_o = np.zeros(l, dtype=gen_specs['out'])
-    H_o['x'][:] = x
-    H_o['consensus_pt'][:] = False
-    H_o['obj_component'][:] = f_i_idxs
-    H_o['get_grad'][:] = True
-    H_o = np.reshape(H_o, newshape=(-1,))      # unfold into 1d array
-
-    tag, Work, calc_in = sendrecv_mgr_worker_msg(libE_info['comm'], H_o)
-
-    gradf_is = calc_in['gradf_i']
-    gradf    = np.sum(gradf_is, axis=0)
-
-    return gradf
-
-def get_neighbor_vals(x, local_gen_id, A_gen_ids_no_local, gen_specs, libE_info):
-    """ Sends local gen data (@x) and retrieves neighbors local data.
-        Sorts the data so the gen ids are in increasing order
-
-    Parameters
-    ----------
-    x : np.ndarray
-        - local input variable
-
-    local_gen_id : int
-        - this gen's gen_id
-
-    A_gen_ids_local : int
-        - expected neighbor's gen ids, not including local gen id
-
-    gen_specs, libE_info : ?
-        - objects to communicate and construct mini History array
-
-    Returns
-    -------
-    X : np.ndarray 
-        - 2D array of neighbors and local x values sorted by gen_ids
-    """
-    H_o = np.zeros(1, dtype=gen_specs['out'])
-    H_o['x'][0] = x
-    H_o['consensus_pt'][0] = True
-
-    tag, Work, calc_in = sendrecv_mgr_worker_msg(libE_info['comm'], H_o)
-
-    neighbor_X = calc_in['x']
-    neighbor_gen_ids = calc_in['gen_worker']
-
-    assert local_gen_id not in neighbor_gen_ids, 'Local data should not be ' + \
-                                                 'sent back from manager'
-    assert np.array_equal(A_gen_ids_no_local, neighbor_gen_ids),'Expected ' + \
-                'gen_ids {}, received {}'.format(A_gen_ids, gen_ids)
-
-    X = np.vstack((neighbor_X, x))
-    gen_ids = np.append(neighbor_gen_ids, local_gen_id)
-
-    # sort data (including local) in corresponding gen_id increasing order
-    X[:] = X[np.argsort(gen_ids)]
-
-    return X
